// frontend/src/backend.d.ts
export interface QuizQuestion {
    question: string;
    correctAnswer: bigint;
    options: Array<string>;
}
// ...
export interface QuestStep {
    id: bigint;
    title: string;
    content: string;
    stepType: string;
    quizQuestions?: Array<QuizQuestion>;
    videoUrl?: string;
}